+++ CHANGELOG.md
@@ -5,6 +5,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.4.16] - 2025-02-18
+
+### Added
+- Added mpv command for media playback
+- Support common mpv options like --start, --length, --volume etc
+
 ## [0.4.15] - 2025-02-18
 
 ### Documentation